third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/gen-device-disconnects-before.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
  return getConnectedHealthThermometerDevice()
    .then(({device, fake_peripheral}) => {
      return disconnectFakeAndWaitForEvent(device, fake_peripheral)
        .then(() => assert_promise_rejects_with_message(
          device.gatt.getPrimaryService('heart_rate'),
          new DOMException(
            'GATT Server is disconnected. ' +
            'Cannot retrieve services. ' +
            '(Re)connect first with `device.gatt.connect`.',
            'NetworkError')));
      });
}, 'Device disconnects before getPrimaryService. Reject with NetworkError.');

</script>
